\name{ADMMRcppArma-package}
\alias{ADMMRcppArma-package}
\alias{ADMMRcppArma}
\docType{package}
\title{
\packageTitle{ADMMRcppArma}
}
\description{
\packageDescription{ADMMRcppArma}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ADMMRcppArma}
\packageIndices{ADMMRcppArma}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ADMMRcppArma}

Maintainer: \packageMaintainer{ADMMRcppArma}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
